Optimization Fake* methods for tests
