+++ inter/event.go
@@ -244,3 +244,34 @@ func FakeFuzzingEvents() (res []*Event) {
 	}
 	return
 }
+
+// FakeEventWithOneEpoch generates random independent events with one epoch for test purpose.
+func FakeEventWithOneEpoch() (res []*Event) {
+	creators := []common.Address{
+		{},
+		hash.FakePeer(),
+		hash.FakePeer(),
+		hash.FakePeer(),
+	}
+	parents := []hash.Events{
+		hash.FakeEventsEpoch(1),
+		hash.FakeEventsEpoch(2),
+		hash.FakeEventsEpoch(8),
+	}
+	i := 0
+	for c := 0; c < len(creators); c++ {
+		for p := 0; p < len(parents); p++ {
+			e := NewEvent()
+			e.Epoch = hash.FakeEpoch()
+			e.Seq = idx.Event(p)
+			e.Creator = creators[c]
+			e.Parents = parents[p]
+			e.Extra = []byte{}
+			e.Sig = []byte{}
+
+			res = append(res, e)
+			i++
+		}
+	}
+	return
+}
